+++ src/main/java/com/felipeleres/customermanagement/dto/ProcessoDTO.java
@@ -4,6 +4,7 @@ import com.felipeleres.customermanagement.entities.Processo;
 import com.felipeleres.customermanagement.enums.FormaPagamento;
 import com.felipeleres.customermanagement.enums.Situacao;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class ProcessoDTO {
@@ -12,7 +13,7 @@ public class ProcessoDTO {
     private String numero;
     private String descricao;
     private LocalDate data;
-    private Double valor;
+    private BigDecimal valor;
     private Situacao situacao;
     private FormaPagamento formaPagamento;
 
@@ -20,7 +21,7 @@ public class ProcessoDTO {
 
     }
 
-    public ProcessoDTO(Long id, String descricao, String numero, LocalDate data, Double valor,Situacao situacao,FormaPagamento formaPagamento) {
+    public ProcessoDTO(Long id, String descricao, String numero, LocalDate data, BigDecimal valor,Situacao situacao,FormaPagamento formaPagamento) {
         this.id = id;
         this.descricao = descricao;
         this.numero = numero;
@@ -72,11 +73,11 @@ public class ProcessoDTO {
         this.descricao = descricao;
     }
 
-    public Double getValor() {
+    public BigDecimal getValor() {
         return valor;
     }
 
-    public void setValor(Double valor) {
+    public void setValor(BigDecimal valor) {
         this.valor = valor;
     }
