+++ src/main/java/com/felipeleres/customermanagement/dto/ProcessoReturnDTO.java
@@ -6,6 +6,7 @@ import com.felipeleres.customermanagement.enums.Situacao;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class ProcessoReturnDTO {
@@ -17,7 +18,7 @@ public class ProcessoReturnDTO {
     private LocalDate data;
     @NotNull(message = "Campo requirido, informe um número de processo!")
     @Positive(message = "O valor deve ser positivo!")
-    private Double valor;
+    private BigDecimal valor;
     private Situacao situacao;
     private ClienteDTO cliente;
     private FormaPagamento formaPagamento;
@@ -26,7 +27,7 @@ public class ProcessoReturnDTO {
 
     }
 
-    public ProcessoReturnDTO(Long id, String descricao, String numero, LocalDate data, Double valor, Situacao situacao, ClienteDTO cliente, FormaPagamento formaPagamento) {
+    public ProcessoReturnDTO(Long id, String descricao, String numero, LocalDate data, BigDecimal valor, Situacao situacao, ClienteDTO cliente, FormaPagamento formaPagamento) {
         this.id = id;
         this.descricao = descricao;
         this.numero = numero;
@@ -80,11 +81,11 @@ public class ProcessoReturnDTO {
         this.descricao = descricao;
     }
 
-    public Double getValor() {
+    public BigDecimal getValor() {
         return valor;
     }
 
-    public void setValor(Double valor) {
+    public void setValor(BigDecimal valor) {
         this.valor = valor;
     }
